KimChang
유산 몇몇
85: 애매함 rewardManager, eventManager, rogueUnitDataBase,
90: 힘듬
96: 힘듬
130: 힘듬
47,48,52 은 못함
제외하고 제작

+++ Assets/KimChang/KimScript/Data/UnitLoader.cs
@@ -48,10 +48,10 @@ public class UnitLoader
             int defaultPrice = Convert.ToInt32(unitData["defaultPrice"]);
             int rarity = Convert.ToInt32(unitData["rarity"]);
             float health = Convert.ToSingle(unitData["health"]);
-            float armor = Convert.ToSingle(unitData["armor"]);
+            int armor = Convert.ToInt32(unitData["armor"]);
             float attackDamage = Convert.ToSingle(unitData["attackDamage"]);
-            float mobility = Convert.ToSingle(unitData["mobility"]);
-            float range = Convert.ToSingle(unitData["range"]);
+            int mobility = Convert.ToInt32(unitData["mobility"]);
+            int range = Convert.ToInt32(unitData["range"]);
             int energy = Convert.ToInt32(unitData["energy"]);
 
             // 불리언 특성
